printserver/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pywin32>=308
Pillow>=11.0.0
python-barcode>=0.15.1
qrcode>=8.0
pyinstaller>=6.0
